feat(divide-number): Add divide two number endpoint. Split code into mulitple files to increase readability.
