+++ main.go
@@ -5,22 +5,12 @@ import (
 	"encoding/json"
 	"github.com/gorilla/mux"
 	"log"
-	"io/ioutil"
-	"fmt"
 )
 
 var people []Person
 
 func GetPeople(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(people)
-}
-
-func AddTwoNumbers(firstNumber float32, secondNumber float32) float32{
-	return firstNumber + secondNumber;
-}
-
-func SquareNumber(number float32) float32{
-	return number * number;
 }
 
 type Person struct {
@@ -34,43 +24,6 @@ type CalcNumbers struct {
 	Number2 float32
 }
 
-// PostHandler converts post request body to string
-func AddTwoNumbersHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == "POST" {
-		body, err := ioutil.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, "Error reading request body",
-				http.StatusInternalServerError)
-		}
-		var calcNumbers CalcNumbers
-		if err := json.Unmarshal([]byte(body), &calcNumbers); err != nil {
-			panic(err)
-		}
-
-		fmt.Fprint(w, AddTwoNumbers(calcNumbers.Number1, calcNumbers.Number2))
-	} else {
-		http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-	}
-}
-
-// PostHandler converts post request body to string
-func SquareHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == "POST" {
-		body, err := ioutil.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, "Error reading request body",
-				http.StatusInternalServerError)
-		}
-		var calcNumbers CalcNumbers
-		if err := json.Unmarshal([]byte(body), &calcNumbers); err != nil {
-			panic(err)
-		}
-
-		fmt.Fprint(w, SquareNumber(calcNumbers.Number1))
-	} else {
-		http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-	}
-}
 
 func provideTestPeople(){
 	people = append(people, Person{ID: "1", Firstname: "Patrick", Lastname: "Schadler"})
@@ -85,6 +38,7 @@ func main() {
 	router.HandleFunc("/people", GetPeople).Methods("GET")
 	router.HandleFunc("/addTwoNumber", AddTwoNumbersHandler)
 	router.HandleFunc("/squareNumber", SquareHandler)
+	router.HandleFunc("/divideNumber", DivideNumbersHandler)
 
 	log.Fatal(http.ListenAndServe(":8000", router))
 }
